backend/pyproject.toml
[project]
name = "hackathon-todo-backend"
version = "0.1.0"
description = "Backend for Hackathon Todo App"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.3",
    "httpx>=0.25.0",
    "opentelemetry-instrumentation>=0.60b1",
    "slowapi>=0.1.9",
    "bleach>=6.0.0",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
